requirements.txt
azure-ai-documentintelligence>=1.0.0b4,<2
azure-storage-blob>=12.19.0,<13
openai>=1.40.0,<2
pypdf>=5.0.0,<6
python-dotenv>=1.0.1,<2
